refactor(billing): split src/lib/billing/actions.ts by domain (CHAOS-1570)
Split the 776-LOC monolithic actions.ts into focused domain modules:

- actions/subscription.ts (203 LOC): getSubscription, startTrialCheckout,
  getSubscriptions, getSubscriptionHistory, changePlan, cancelSubscription,
  reactivateSubscription, getBillingPortalUrl
- actions/invoices.ts (163 LOC): getInvoices, getInvoice, voidInvoice
- actions/refunds.ts (130 LOC): createRefund, getRefunds
- actions/checkout.ts (208 LOC): listBillingPlans, createBillingPlan,
  updateBillingPlan, deleteBillingPlan, syncBillingPlanToStripe, pullPlansFromStripe
- actions/_shared.ts (95 LOC): private helpers (sanitizeId, getAuthHeaders,
  resolveOrgId, getBackendUrl, withErrorHandling, apiRequest)

Original actions.ts becomes a 6-line barrel re-exporting all domain modules.
All 16 import sites unchanged — existing consumers keep working without modification.
Each domain file starts with 'use server' directive on line 1.
